Convert AuRa block producer wiring to DI (#12414)
* refactor(aura): convert AuRa block producer wiring to DI

Replace the hand-built StartBlockProducerAuRa factory with the standard
block producer env pattern:

- AuRaTxPoolTxSourceFactory: extracted tx pool source creation
  (tx priority contract / local data source), also bound as the
  IBlockProducerTxSourceFactory for the post-merge env in Merge.AuRa
- AuRaBlockProducerEnvFactory: producer env scope registering the
  sealing IAuRaValidator (shared by processor and tx source), producer
  tx filter, gas limit override, contract rewriter and an ITxSource
  decorator adding posdao/randomness/generated tx sources
- AuRaBlockProducerFactory now uses IBlockProducerEnvFactory
  CreatePersistent like other consensus plugins

Adds a container wiring regression test initializing the producer and
runner from a full AuRa container.


* test(aura): cover pre/post-merge block producer wiring in merge container

Verifies that with AuRaModule + AuRaMergeModule the decorated
IBlockProducerFactory builds a MergeBlockProducer whose pre-merge
producer resolves through AuRaBlockProducerEnvFactory (including the
scoped AuRaMergeBlockProcessor override) and that the runner factory
chain produces a MergeBlockProducerRunner.


* refactor(aura): keep original CheckAdd* tx source structure for reviewability

Restores the CheckAddPosdaoTransactions/CheckAddRandomnessTransactions
local functions and needSigner |= composition verbatim from the removed
StartBlockProducerAuRa.CreateTxSourceForProducer.


* test(aura): drop manual registrations covered by AuRaMergeModule

MergeAuRaTestBlockchain manually registered the withdrawal components
and NullSigner that AuRaMergeModule (added right above) and the base
TestBlockProcessingModule already provide.


* test(aura): restore merge-container block producer wiring test

Re-adds AuRaMergeBlockProducerWiringTests so the pre-merge
AuRaBlockProducerEnvFactory resolution inside a merge container
(scoped AuRaMergeBlockProcessor override active) stays regression-guarded.


---------
